requirements.txt
python-dotenv==1.0.1
llama-index==0.12.23
llama-index-core==0.12.23.post2
google-generativeai==0.8.4
llama-index-llms-openai==0.3.25
llama-index-llms-gemini==0.4.11
llama-index-llms-anthropic==0.6.10
fastapi==0.112.2
colorama==0.4.6
streamlit==1.39.0
